+++ src/index.ts
@@ -1,7 +1,11 @@
-import Button from "./Button/Button";
-import Input from "./Input/Input";
+import Button from "./components/Button/Button";
+import Input from "./components/Input/Input";
+import { AquaSelect } from "./components/Select/Select";
+import Panel from "./components/Panel/Panel";
 
 export {
-    Button,
-    Input
+  Button,
+  Input,
+  AquaSelect as Select,
+  Panel
 }