+++ changelog.md
@@ -1,3 +1,5 @@
 ## 0.1.0
 
 * Initial stages of project
+* Create components: Button, Input, Panel
+* Add Storybook and docs